refactor: Add TargetedYearlyBackup command for creating yearly backups of specific tables
The `TargetedYearlyBackup` command is added to the `BackupDatabase.php` file. This command allows users to create a yearly backup of specific tables in the database. The backup is created by copying the structure and data of the specified tables to a new database with the current year appended to its name. The original tables are then cleared of data. This feature improves data management and provides a convenient way to perform yearly backups.