refactor: Add new view for printing type 4 reports
A new view 'print4.blade.php' was added to handle the printing of type 4 reports. This view is used when the 'jenis_laporan' input is set to 4 in the LaporanController's print method. It includes the necessary data for generating the report. This change improves the organization and separation of concerns in the codebase.