refactor: Improve organization and separation of concerns in print4.blade.php
Refactor the print4.blade.php view to improve the organization and separation of concerns. Instead of using multiple if-else statements to determine the background color of table rows based on the 'roda_kendaraan' value, use an associative array to map the 'roda_kendaraan' values to their corresponding group index. This change simplifies the code and makes it more maintainable.